Refactor background.js: Move Codewars to Github logic to separate file
